feat: signup and login

+++ backend/src/services/recipes.js
@@ -1,7 +1,8 @@
 import { Recipe } from '../db/models/recipe.js'
+import { User } from '../db/models/user.js'
 
-export async function createRecipe({ title, ingredients, image }) {
-  const recipe = new Recipe({ title, ingredients, image })
+export async function createRecipe(userId, { title, ingredients, image }) {
+  const recipe = new Recipe({ title, author: userId, ingredients, image })
   return await recipe.save()
 }
 
@@ -16,6 +17,11 @@ async function listRecipes(
 export async function listAllRecipes(options) {
   return await listRecipes({}, options)
 }
+export async function listRecipesByAuthor(authorUsername, options) {
+  const user = await User.findOne({ username: authorUsername })
+  if (!user) return []
+  return await listRecipes({ author: user._id }, options)
+}
 // list posts by title by passing in title as query to listRecipes()
 export async function listRecipesByTitle(title, options) {
   return await listRecipes({ title }, options)
@@ -28,15 +34,19 @@ export async function listRecipesByIngredient(ingredients, options) {
 export async function getRecipeById(recipeId) {
   return await Recipe.findById(recipeId)
 }
-// update a post
-export async function updateRecipe(recipeId, { title, ingredients, image }) {
+// update a recipe
+export async function updateRecipe(
+  userId,
+  recipeId,
+  { title, ingredients, image },
+) {
   return await Recipe.findOneAndUpdate(
-    { _id: recipeId },
+    { _id: recipeId, author: userId },
     { $set: { title, ingredients, image } },
     { new: true },
   )
 }
-// delete a post
-export async function deleteRecipe(recipeId) {
-  return await Recipe.deleteOne({ _id: recipeId })
+// delete a recipe
+export async function deleteRecipe(userId, recipeId) {
+  return await Recipe.deleteOne({ _id: recipeId, author: userId })
 }
